Refactoring state management
