+++ packages/app/src/components/presentational/MyFyghterContainer.tsx
@@ -22,7 +22,7 @@ export const MyFyghterContainer: React.FC = () => {
     };
     init();
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [myFyghter, account]);
+  }, [account]);
 
   const hasFyghter = !isLoading && myFyghter !== null;
